Refactor: 타입 시스템 완전 통합 및 타입 에러 해결 - API, 훅, 컴포넌트 타입을 src/types/로 완전 통합 - src/types/api/client.ts, src/types/hooks/ 폴더 생성 - useApi 타입 추론 문제 해결 및 AuthContext 타입 에러 수정 - 중복 타입 정의 제거 및 import 경로 정리 - 타입 시스템 일관성 및 재사용성 대폭 향상
